Added separate folder for login-registration, added email and role state to app.js and accordingly moddified login and registration
