+++ libraryDB/frontend/src/App.js
@@ -1,16 +1,20 @@
-import { useContext } from "react";
+import { useContext, useState } from "react";
 import { Route, Switch } from 'react-router-dom';
-import Login from './Login'
+import Login from './login-registration/Login'
 import Register from './Register'
 import {roleContext} from "./RoleContext"
 
 export default function App() {
 
+  const [role, setRole] = useState('');
+  const [email, setEmail] = useState('');
+
+
   return (
     <div className="App">
       <Switch>
           <Route path='/login'>
-            <Login/>
+            <Login setRole={setRole} setEmail={setEmail}/>
           </Route>
           <Route path='/register'>
             <Register/>
