+++ libraryDB/frontend/src/login-registration/Login.js
@@ -1,16 +1,16 @@
 import {useState} from 'react';
 import {Grid, TextField, Button, Tab, Typography} from "@material-ui/core"
 import {useHistory, Link} from "react-router-dom";
-import {roles} from "./Roles"
-import {roleContext} from "./RoleContext"
 
-export default function Login() {
+export default function Login(props) {
 
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
-    const history = useHistory();
 
-    let userRole;
+    let {setRole} = props;
+    let {setEmail} = props;
+
+    const history = useHistory();
 
     function handleLogin() {
         fetch(`/api/account?email=${email}&password=${password}`)
@@ -20,14 +20,15 @@ export default function Login() {
                         response.json()
                             .then(account =>{
                                userRole = account.hasOwnProperty("role") ? account["role"] : "user";
+                               setRole(userRole);
+                               setEmail(account["email"]);
+                               history.push(`/${userRole}`);
                             })
                     }
                     else {
                         alert("Incorrect credentials");
                     }
                     return response.ok;
-            }).catch(error =>{
-                alert("Incorrect credentials");
             })
     }
 
@@ -46,20 +47,16 @@ export default function Login() {
                     <Grid item xs={12}>
                         <TextField id="password-input" label="password" onInput={ e =>setPassword(e.target.value)} />
                     </Grid>
-                    <roleContext.Provider value={userRole}>
                         <Grid item xs={12}>
                             <Link to={handleLogin() ? '/' : '#'}>  
                                 <Button type="submit" variant="contained" size="large" color="primary">Login</Button>
                             </Link>  
                         </Grid>
-                    </roleContext.Provider>
-                    <roleContext.Provider value={roles.USER}>
                         <Grid item xs={12}>
                             <Typography>
                                 <Tab label='Register'  to='/register' component={Link} />
                             </Typography>
                         </Grid>
-                    </roleContext.Provider>
                 </form>
             </div>
         </Grid>
