+++ libraryDB/frontend/src/login-registration/Register.js
@@ -11,10 +11,8 @@ export default function Register() {
     const [surname, setSurname] = useState('');
     const [email, setEmail] = useState('');
     const [password, setPassword] = useState('');
-    const [role, setRole] = useState('');
 
     const history = useHistory();
-    const userRole = useContext(roleContext);
 
     function handleRegistration(event) {
         event.preventDefault();
@@ -26,12 +24,6 @@ export default function Register() {
             password: password,
         }
 
-                    console.log(userRole);
-
-        if(userRole.normalize() === roles.ADMIN.normalize()){
-            requestBody.role = role;
-        }
-
         fetch('/api/user', {
             method: "POST",
             headers: {
@@ -41,40 +33,14 @@ export default function Register() {
             body: JSON.stringify(requestBody)
         }).then(response =>{
             if(!response.ok){
-                let okRegistrationReturnPath = userRole.normalize() === roles.USER ? '/login' : '/'; 
-                history.push(okRegistrationReturnPath);
+                alert("Account created succesfully");
+                history.push('/login');
+            }
+            else{
+                history.push('#');
             }
         })
     }
-
-    function handleRoleMenuItemChange(event) {
-        setRole(event.target.value);
-      };
-
-      let AdminRolePanel = null;
-      if(userRole.normalize() === roles.ADMIN.normalize()){
-        AdminRolePanel = (
-            <Grid item xs={12}>
-            <FormControl>
-                <InputLabel id="role-select">Role</InputLabel>
-                <Select
-                    labelId="role-select"
-                    id="role-select"
-                    value={role}
-                    onChange={handleRoleMenuItemChange}
-                    >
-                    {() =>{
-                        const menuItems = [];
-                        for (const [key, value] of Object.entries(roles)) {
-                            menuItems.push(<MenuItem value={value}>{key}</MenuItem>);
-                          }
-                          return menuItems;
-                    }}
-                </Select>
-            </FormControl>
-        </Grid>
-        );
-      }
 
     return(
         <Grid
@@ -97,7 +63,6 @@ export default function Register() {
                     <Grid item xs={12}>
                         <TextField id="password-input" label="password" onInput={ e =>setPassword(e.target.value)} />
                     </Grid>
-                    {AdminRolePanel}
                     <Grid item xs={12}>
                         <Button type="submit" variant="contained" size="large" color="primary">Register</Button>
                     </Grid>
